requirements.txt
SQLAlchemy==2.0.27
python-telegram-bot==20.8
python-dotenv==1.0.1 